Read ME
Ladybird Flower Power Code

Ladybird the flower power style The ladybird flower power, gives a bright and cute visualisation of nature and technology link. Creating a fun and adventure code of exploring different colour combinations and drawing a way through the code to discover new features. Allowing for the user to undertake a creative and fun method of interaction with the code.